backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
sqlalchemy
pydantic
aiofiles
pdfplumber
python-docx
langchain
openai
requests
numpy
